Move default resampling function
The default resampling function, `average`, is moved from the actor to
the `timeseries.resampling` module so the `Resampler` class can have a
default too.
